+++ src/raftstore/store/mod.rs
@@ -42,8 +42,9 @@ pub use self::engine::{Iterable, Mutable, Peekable};
 pub use self::peer_storage::{do_snapshot, init_apply_state, init_raft_state, write_peer_state,
                              CacheQueryStats, PeerStorage, SnapState, RAFT_INIT_LOG_INDEX,
                              RAFT_INIT_LOG_TERM};
-pub use self::snap::{check_abort, copy_snapshot, ApplyOptions, SnapEntry, SnapKey, SnapManager,
-                     SnapManagerBuilder, Snapshot, SnapshotDeleter, SnapshotStatistics};
+pub use self::snap::{check_abort, copy_snapshot, ApplyOptions, Error as SnapError, SnapEntry,
+                     SnapKey, SnapManager, SnapManagerBuilder, Snapshot, SnapshotDeleter,
+                     SnapshotStatistics};
 pub use self::region_snapshot::{RegionIterator, RegionSnapshot};
 
 // Only used in tests
